// src/api/types.gen.ts
export type CodeSnippetRowElementModel = IApiElementModelBase & {
    properties?: CodeSnippetRowPropertiesModel;
};

export type CodeSnippetRowPropertiesModel = {
    title?: (string) | null;
    code?: (string) | null;
};

export type CodeSnippetRowSettingsElementModel = IApiElementModelBase & {
    properties?: CodeSnippetRowSettingsPropertiesModel;
};

export type CodeSnippetRowSettingsPropertiesModel = HidePropertyPropertiesModel & SpacingPropertiesPropertiesModel;
// ...
export type HidePropertyElementModel = IApiElementModelBase & {
    properties?: HidePropertyPropertiesModel;
};

export type HidePropertyPropertiesModel = {
    hide?: (boolean) | null;
};
// ...
export type IApiContentModel = ArticleControlsContentModel | ContentControlsContentModel | HeaderControlsContentModel | MainImageControlsContentModel | SEocontrolsContentModel | VisibilityControlsContentModel | ArticleContentModel | ArticleListContentModel | AuthorContentModel | AuthorListContentModel | CategoryContentModel | CategoryListContentModel | ContactFormControlsContentModel | ContactContentModel | ContentContentModel | ErrorContentModel | FooterControlsContentModel | HomeContentModel | SearchContentModel | XMlsitemapContentModel | LoginContentModel;
// ...
export type IApiElementModel = ListPageSettingsElementModel | CodeSnippetRowElementModel | HidePropertyElementModel | SpacingPropertiesElementModel | CodeSnippetRowSettingsElementModel | IconLinkRowElementModel | IconLinkRowSettingsElementModel | ImageCarouselRowElementModel | ImageCarouselRowSettingsElementModel | ImageRowElementModel | ImageRowSettingsElementModel | LatestArticlesRowSettingsElementModel | LatestArticlesRowElementModel | RichTextRowElementModel | RichTextRowSettingsElementModel | VideoRowElementModel | VideoRowSettingsElementModel;

export type IApiElementModelBase = {
    readonly id?: string;
    readonly contentType: string;
};
// ...
export type IconLinkRowElementModel = IApiElementModelBase & {
    properties?: IconLinkRowPropertiesModel;
};

export type IconLinkRowPropertiesModel = {
    icon?: Array<IApiMediaWithCropsModel> | null;
    link?: Array<ApiLinkModel> | null;
};

export type IconLinkRowSettingsElementModel = IApiElementModelBase & {
    properties?: IconLinkRowSettingsPropertiesModel;
};

export type IconLinkRowSettingsPropertiesModel = HidePropertyPropertiesModel;

export type ImageCarouselRowElementModel = IApiElementModelBase & {
    properties?: ImageCarouselRowPropertiesModel;
};

export type ImageCarouselRowPropertiesModel = {
    images?: Array<IApiMediaWithCropsModel> | null;
};

export type ImageCarouselRowSettingsElementModel = IApiElementModelBase & {
    properties?: ImageCarouselRowSettingsPropertiesModel;
};

export type ImageCarouselRowSettingsPropertiesModel = HidePropertyPropertiesModel & SpacingPropertiesPropertiesModel;

export type ImageCropCoordinatesModel = {
    x1?: number;
    y1?: number;
    x2?: number;
    y2?: number;
};

export type ImageCropModel = {
    alias?: (string) | null;
    width?: number;
    height?: number;
    coordinates?: ImageCropCoordinatesModel;
};

export type ImageFocalPointModel = {
    left?: number;
    top?: number;
};

export type ImageRowElementModel = IApiElementModelBase & {
    properties?: ImageRowPropertiesModel;
};

export type ImageRowPropertiesModel = {
    image?: Array<IApiMediaWithCropsModel> | null;
    caption?: (string) | null;
};

export type ImageRowSettingsElementModel = IApiElementModelBase & {
    properties?: ImageRowSettingsPropertiesModel;
};

export type ImageRowSettingsPropertiesModel = SpacingPropertiesPropertiesModel & HidePropertyPropertiesModel;

export type LatestArticlesRowElementModel = IApiElementModelBase & {
    properties?: LatestArticlesRowPropertiesModel;
};

export type LatestArticlesRowPropertiesModel = {
    articleList?: IApiContentModel;
    pageSize?: (number) | null;
    showPagination?: (boolean) | null;
};

export type LatestArticlesRowSettingsElementModel = IApiElementModelBase & {
    properties?: LatestArticlesRowSettingsPropertiesModel;
};

export type LatestArticlesRowSettingsPropertiesModel = HidePropertyPropertiesModel & SpacingPropertiesPropertiesModel;

export enum LinkTypeModel {
    CONTENT = 'Content',
    MEDIA = 'Media',
    EXTERNAL = 'External'
}

export type ListPageSettingsElementModel = IApiElementModelBase & {
    properties?: ListPageSettingsPropertiesModel;
};

export type ListPageSettingsPropertiesModel = {
    [key: string]: unknown;
};
// ...
export type RichTextRowElementModel = IApiElementModelBase & {
    properties?: RichTextRowPropertiesModel;
};

export type RichTextRowPropertiesModel = {
    content?: RichTextModel;
};

export type RichTextRowSettingsElementModel = IApiElementModelBase & {
    properties?: RichTextRowSettingsPropertiesModel;
};

export type RichTextRowSettingsPropertiesModel = SpacingPropertiesPropertiesModel & HidePropertyPropertiesModel;
// ...
export type SpacingPropertiesElementModel = IApiElementModelBase & {
    properties?: SpacingPropertiesPropertiesModel;
};

export type SpacingPropertiesPropertiesModel = {
    paddingTop?: (string) | null;
    paddingBottom?: (string) | null;
    paddingLeft?: (string) | null;
    paddingRight?: (string) | null;
    marginTop?: (string) | null;
    marginBottom?: (string) | null;
    marginLeft?: (string) | null;
    marginRight?: (string) | null;
};

export type VideoRowElementModel = IApiElementModelBase & {
    properties?: VideoRowPropertiesModel;
};

export type VideoRowPropertiesModel = {
    videoUrl?: (string) | null;
    caption?: (string) | null;
};

export type VideoRowSettingsElementModel = IApiElementModelBase & {
    properties?: VideoRowSettingsPropertiesModel;
};

export type VideoRowSettingsPropertiesModel = HidePropertyPropertiesModel & SpacingPropertiesPropertiesModel;